fix: unwind fixture, seemed to not work in CI for some reason
