+++ examples/_deprecated/concurrency_limit_rr/test_dep_concurrency_limit_rr.py
@@ -5,10 +5,12 @@ import pytest
 
 from hatchet_sdk import Hatchet
 from hatchet_sdk.workflow_run import WorkflowRunRef
+from tests.utils import fixture_bg_worker
+
+worker = fixture_bg_worker(["poetry", "run", "concurrency_limit_rr"])
 
 
 # requires scope module or higher for shared event loop
-@pytest.mark.parametrize("worker", ["concurrency_limit_rr"], indirect=True)
 @pytest.mark.skip(reason="The timing for this test is not reliable")
 @pytest.mark.asyncio(scope="session")
 async def test_run(aiohatchet: Hatchet, worker):
