+++ examples/bulk_fanout/test_bulk_fanout.py
@@ -1,11 +1,13 @@
 import pytest
 
 from hatchet_sdk import Hatchet
+from tests.utils import fixture_bg_worker
+
+worker = fixture_bg_worker(["poetry", "run", "bulk_fanout"])
 
 
 # requires scope module or higher for shared event loop
 @pytest.mark.asyncio(scope="session")
-@pytest.mark.parametrize("worker", ["bulk_fanout"], indirect=True)
 async def test_run(hatchet: Hatchet, worker):
     run = hatchet.admin.run_workflow("BulkParent", {"n": 12})
     result = await run.result()
@@ -14,7 +16,6 @@ async def test_run(hatchet: Hatchet, worker):
 
 # requires scope module or higher for shared event loop
 @pytest.mark.asyncio(scope="session")
-@pytest.mark.parametrize("worker", ["bulk_fanout"], indirect=True)
 async def test_run2(hatchet: Hatchet, worker):
     run = hatchet.admin.run_workflow("BulkParent", {"n": 10})
     result = await run.result()
