+++ examples/cancellation/test_cancellation.py
@@ -1,11 +1,13 @@
 import pytest
 
 from hatchet_sdk import Hatchet
+from tests.utils import fixture_bg_worker
+
+worker = fixture_bg_worker(["poetry", "run", "cancellation"])
 
 
 # requires scope module or higher for shared event loop
 @pytest.mark.asyncio(scope="session")
-@pytest.mark.parametrize("worker", ["cancellation"], indirect=True)
 async def test_run(hatchet: Hatchet, worker):
     run = hatchet.admin.run_workflow("CancelWorkflow", {})
     result = await run.result()
