+++ examples/concurrency_limit/test_concurrency_limit.py
@@ -2,12 +2,14 @@ import pytest
 
 from hatchet_sdk import Hatchet
 from hatchet_sdk.workflow_run import WorkflowRunRef
+from tests.utils import fixture_bg_worker
+
+worker = fixture_bg_worker(["poetry", "run", "concurrency_limit"])
 
 
 # requires scope module or higher for shared event loop
-@pytest.mark.asyncio(scope="session")
 @pytest.mark.skip(reason="The timing for this test is not reliable")
-@pytest.mark.parametrize("worker", ["concurrency_limit"], indirect=True)
+@pytest.mark.asyncio(scope="session")
 async def test_run(hatchet: Hatchet, worker):
     num_runs = 6
     runs: list[WorkflowRunRef] = []
