+++ examples/fanout/test_fanout.py
@@ -1,11 +1,13 @@
 import pytest
 
 from hatchet_sdk import Hatchet
+from tests.utils import fixture_bg_worker
+
+worker = fixture_bg_worker(["poetry", "run", "fanout"])
 
 
 # requires scope module or higher for shared event loop
 @pytest.mark.asyncio(scope="session")
-@pytest.mark.parametrize("worker", ["fanout"], indirect=True)
 async def test_run(hatchet: Hatchet, worker):
     run = hatchet.admin.run_workflow("Parent", {"n": 2})
     result = await run.result()
@@ -14,7 +16,6 @@ async def test_run(hatchet: Hatchet, worker):
 
 # requires scope module or higher for shared event loop
 @pytest.mark.asyncio(scope="session")
-@pytest.mark.parametrize("worker", ["fanout"], indirect=True)
 async def test_run2(hatchet: Hatchet, worker):
     run = hatchet.admin.run_workflow("Parent", {"n": 2})
     result = await run.result()
