+++ examples/logger/test_logger.py
@@ -1,11 +1,13 @@
 import pytest
 
 from hatchet_sdk import Hatchet
+from tests.utils import fixture_bg_worker
+
+worker = fixture_bg_worker(["poetry", "run", "logger"])
 
 
 # requires scope module or higher for shared event loop
 @pytest.mark.asyncio(scope="session")
-@pytest.mark.parametrize("worker", ["logger"], indirect=True)
 async def test_run(hatchet: Hatchet, worker):
     run = hatchet.admin.run_workflow("LoggingWorkflow", {})
     result = await run.result()
