+++ examples/timeout/test_timeout.py
@@ -1,11 +1,13 @@
 import pytest
 
 from hatchet_sdk import Hatchet
+from tests.utils import fixture_bg_worker
+
+worker = fixture_bg_worker(["poetry", "run", "timeout"])
 
 
 # requires scope module or higher for shared event loop
 @pytest.mark.asyncio(scope="session")
-@pytest.mark.parametrize("worker", ["timeout"], indirect=True)
 async def test_run_timeout(hatchet: Hatchet, worker):
     run = hatchet.admin.run_workflow("TimeoutWorkflow", {})
     try:
@@ -16,7 +18,6 @@ async def test_run_timeout(hatchet: Hatchet, worker):
 
 
 @pytest.mark.asyncio(scope="session")
-@pytest.mark.parametrize("worker", ["timeout"], indirect=True)
 async def test_run_refresh_timeout(hatchet: Hatchet, worker):
     run = hatchet.admin.run_workflow("RefreshTimeoutWorkflow", {})
     result = await run.result()
