+++ tests/utils/__init__.py
@@ -1,0 +1,1 @@
+from .bg_worker import fixture_bg_worker
